Committing the package "multipaxos", a first cut implementation of paxos that has been only minimally tested. It is not full featured and does not yet include support for logging and stop requests (needed for group changes). But I am committing this now as a checkpoint so that the existing and planned interfaces between "nameserver" and "replicaCoordination" can converge smoothly in small steps.
The multipaxos code is currently inactive as a standalone package, so
it should not break anything in the current GNS.

The classes PaxosManager and PaxosAcceptorState can be tested by just
running them. The latter test shows that an inactive paxos instance
consumes less than 170 bytes of space, so we can maintain paxos
instances for tens of millions of GUIDs with a few gigs of memory.

The classes WaitForUtility, PaxosMessenger, and MessagingTask, and
util/NullIfEmpty may be independently useful at other places in GNS.



git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@423 0b15c3b6-4604-4ca4-ad59-2e1af984e93c