MOB-150  In-progress refactoring NS code into : ActiveReplica, ReplicaController, ActiveReplicaCoordination, and ReplicaControllerCoordination.
Interface changes to three classes in GNS - NameRecord, ReplicaControllerRecord, and HashFunction - that were needed due to MOB-150.

Both ReplicaControllerRecord and NameRecord now do not depend on the NameServer class. 
To do so: 
(1) I have moved static methods  related to  ReplicaControllerRecord that were in NameServer to the ReplicaControllerRecord class. Likewise for NameRecord.
(2) All constructors and  methods that read/wr from database now take a parameter BasicRecordMap.

The class HashFunction in util package is changed to ConsistentHashing. The earlier method getPrimaryReplicas is replaced by getReplicaControllerSet. 


git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@455 0b15c3b6-4604-4ca4-ad59-2e1af984e93c