Having a single thread processing incoming messages in JSONMessageExtractor (formerly JSONMessageWorker) had one issue. If that thread were to send a message and call wait() to wait for a response there would be a deadlock.
Implemented a simple enhancement which uses a thread pool so we can handle messages concurrently. Theoretically we could still have a problem
if we limit the pool size and every thread is waiting on a message. 

All of this was in service of support for DNS lookups form an NS back to the LNS during message processing. This particular case involves looking up GUID info
on other servers while doing authentication. Still need to implement writes for this using a similar mechanism before we can finish implementing MOB-180. 

All functionality tests pass on the 3 node local server.
Issue #MOB-180 - Refreshing context-aware group GUIDs

git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@589 0b15c3b6-4604-4ca4-ad59-2e1af984e93c