A hashmap based on multiple arrays (cuckoo hashing) that dramatically reduces the number of additional bytes to store a map of paxos instances from about 300B in HashMap to just under 6B. Needed because Java HashMap takes more bytes for an entry than the size of the paxos instance itself.
MultiArrayMap is generic and can take any KeyType and ValueType
provided the latter implements a getKey() function that returns
KeyType, i.e., it uses getKey() as defined (instead of equals() in
HashMap) to decide if two ValueType values are the same. This means
that the key is part of the value (as opposed to the two being
arbitrary, unrelated objects in HashMap), which is how hashmaps are
often used anyway.



git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@598 0b15c3b6-4604-4ca4-ad59-2e1af984e93c