Committing multipaxos changes with the following feature and testing add-ons.
1) Supports batched logging with connection pooling and yields a write
throughput of about ~4K/sec and a latency of ~10ms under light load
across a LAN.

2) "megapaxos": Idle paxos instances use under 250B of memory, so we
can support up to 4M instances per GB, but the GC kicks in at around
3M instances, so it is safer to assume around 400B/instance.

3) "gigapaxos": Added support for "hibernate-restore" and
"pause-hotRestore". The former checkpoints idle instances and recovers
when a new request is received. The latter is lighter-weight and just
stores an idle paxos instance's meta-info to disk and quickly reads it
back in when a new request comes in. Either of these can potentially
allow us to scale to billions of paxos instances on a commodity
machine in exchange for a small (<10ms) "speed bump" when the first
request in a burst of requests for a record arrives.

4) Added the following testing classes:
TESTPaxosMain: for a single node test
TESTPaxosNode: for a distributed test
TESTPaxosClient: for a distributed test
TESTPaxosReplicable: the default "app" used in tests
TESTPaxosConfig: for configuring test parameters. 

PaxosManager, PaxosCoordinatorState, and PaxosAcceptor can still be
run by themselves for a unit test.

5) Added some utility classes in and moved some existing utility
classes to paxosutil/.  

6) Added retransmission with exponential backoff in Messenger to try
to tolerate NIO packet drops because of intermittent congestion. Added
support for resending prepare and accept messages after a long timeout
as otherwise losses under intermittent congestion can cause the state
machine to get stuck until it is restarted.

7) Added untested support for figuring out your group if you happen to
get a paxos packet for an unfindable instance. This can happen if an
instance got created when a node was previously crashed.

8) Many miscellaneous bug fixes. But there seems to be no dearth of
hidden ones rearing their ugly heads up.



git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@661 0b15c3b6-4604-4ca4-ad59-2e1af984e93c