FIXME:
(1) GNSNodeConfig insists on outdated fake ping latencies and throws
an error. 

(2) GNSNodeConfig or any other place in the system should not have any
support for local nameserver IDs. A local name server is an
independent entity that only receives responses. Nobody proactively
should initiate communication with it. For making this work, packets
should simply include a local name server IP address. It is
straightforward expose IP-address-based messaging in GNSNIOTransport
just as in NIOTransport.

Local name servers are not part of the GNS, so they can not be
expected to have unique IDs.



git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@779 0b15c3b6-4604-4ca4-ad59-2e1af984e93c