Added sendToAddress to GNSNIOTransportInterface. Implemented sendToAddress methods for GNSNIOTransport and PacketTypeStamper. Still need a sendToAddress method for NioServer and GnsMessenger.
Added an LNS class CommandRequest which handles packets coming in from the client (shares code
with the HTTPServer). Also CommandPacket and CommandValueReturnPacket (for lack of a better name).

Writing a test class in the GNS to verify server side fnctionality. 

Still need to write a client-side class that uses Nio to send the original request.

All tests pass except for issues related to MOB-265
Issue #MOB-258 - TCP interface for querying GNS through the client library

git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@792 0b15c3b6-4604-4ca4-ad59-2e1af984e93c