(1) Changed PAXOS_ID_SIZE (same as GUID) from 20 to 40 because the byte representation in hex seems to bloat it up. Not sure about this.
(2) Commented static isFirstConnect() check in line 780 because that
causes the second PaxosManager's database to not be created (an error
Abhigyan ran into). The test cases assumed a single database location
(with different table names for different nodes), so the use case of
multiple paxos managers per node with different DB locations was never
tested until now. 



git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@855 0b15c3b6-4604-4ca4-ad59-2e1af984e93c