MOB-253 Documentation on how to run a simple single-node and multi-node experiment
Added a stand-alone NIO-based client, called DBClient, which can be used to send add/remove/read/write requests to GNS.
The functionality is implemented in package edu.umass.cs.gns.test.nioclient, and more specifically in the classes DBClientIntercessor and DBClient.

The DBClient provides and interface to send requests. There is a sample program called 'ClientSample' written using this interface in the same package.

A step-by-step instruction to run a test using the DBClient and a local GNS installation are provided in GNS_TRUNK/scripts/basic_test_local. 
That folder also has shell scripts to run these tests in 1 step.


git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@863 0b15c3b6-4604-4ca4-ad59-2e1af984e93c