Fixed some minor bugs that probably never showed up before in JSONMessenger.
Added some currently unused code to BasicPacketDemultiplexer. This is
a change I am proposing we should implement. It requires every
inheritor of BasicPacketDemultiplexer to register all PacketTypes it
is designed to handle. The motivation is that BasicPacketDemultiplexer
needs to be multithreaded, and this concern can not be left to the
inheritors.


Note to Abhigyan and Westy: 

I can not think of a better way to do this than requiring inheritors
to register all of their packet types upfront. The alternative would
be to have them implement a "boolean isTypeHandleable(PacketType)"
method. We need this information so that NIO knows whether it is done
handling the packet even though it is scheduled in a separate thread
so that demultiplexer chaining works correctly.

Can you confirm you can implement this change in your respective
inheritors or comment on potential issues?



git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@875 0b15c3b6-4604-4ca4-ad59-2e1af984e93c