MOB-194 Application.handleDecision should return boolean, otherwise the code is incorrect
I committed an update that fixes this issue as follows.
 
I defined  a 'FailedDBOperationException' which replaces the earlier 'FailedUpdateException'.  FailedDBOperationException is thrown upon all types of database operation, e. g., update, lookup, insertion, iteration.

I found that we needed changed not only to return value of Application.handleDecision() but also for return value of other methods in Replicable, and Reconfigurable interfaces. Following are the main changes to code: 

(1)  If a database operation is happens as a result of call to 'Application.handleDecision' (which is the most common case), and we encounter a FailedDBOperationException, then 'Application.handleDecision' returns false. We are assuming that Application.handleDecision() will be called again if return value is 'false'.  As an example, the method handleDecision() in class GnsReconfigurable returns false, and GnsCoordinatorPaxos which calls handleDecision() retries the operation if handleDecision() returns false.

(2)  If 'Replicable.getState()' sees a FailedDBOperationException, we return a null value, which indicates that state cannot be read currently. For example, see GnsReconfigurable.getState().

(3)  If 'Replicable.updateState()' sees a FailedDBOperationException, we return 'false' meaning that the updated state wasn't applied. For example, see GnsReconfigurable.updateState().

(4) If 'Reconfigurable.putInitialState()' sees a FailedDBOperationException, we retry the operation locally. For example, see GnsReconfigurable.putInitialState().

(5) If 'Reconfigurable.stopVersion()' sees a FailedDBOperationException, we return 'false' and hope that this method will be called again. See. GnsReconfigurable.stopVersion()

(6) If 'Reconfigurable.getFinalState()' sees a FailedDBOperationException,  we return a null value indicating that the final state isn't currently available.



git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@908 0b15c3b6-4604-4ca4-ad59-2e1af984e93c