Fixed a bug pointed out by Abhigyan that caused a paxos checkpoint to be logged after the instance is removed. The reason was that checkpointing is an asynchronous operation handed over to a helper task that writes to the database, so it was possible for the following sequence of events to happen: a checkpoint is issued, the paxos instance is killed, the checkpoint completes. This bug leaves the checkpoint hanging without a paxos instance, which effectively results in a "delete GUID" action not actually happening, i.e., if the GUID is re-created, it starts with the old state as opposed to the default initial state.
The fix introduces a kill(.) method in AbstractPaxosLogger that
removes the (only) pending checkpoint for a paxos instance in addition
to removing the checkpoint stored in the database if any.

Also made some minor edits to (1) explicitly verify that the
(reconfiguration) version is verified as matching when a paxos
instance is created, though this ought to be true by design anyway;
(2) change a severe log message about "trapped" paxos messages to a
warning instead, as this event can happen and is not fatal.



git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@918 0b15c3b6-4604-4ca4-ad59-2e1af984e93c