Restored compilation with some hacks. Search for "FIXME NodeId" for where all the hacks need to be fixed.
Overall, the class NodeId needs to go. Most of the FIXME's
above are places where a generic type is being converted to 
NodeId<String> (or a Set thereof) for no good reason.

multipaxos.PaxosManager will currently work if provided Integer node
IDs. But the rest of the code will probably break until forced
conversions to NodeId are removed.



git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/branches/v1.11-2014-9-22-arun-merge@1139 0b15c3b6-4604-4ca4-ad59-2e1af984e93c