Changed PaxosManager.propose for backwards compatibility with Auspice. gigapaxos will now return a String to app.handleDecision(.) if a String request is proposed and will return a RequestPacket.toString() if a RequestPacket is proposed. But this is still a bit different from the older implementation that always returns the RequestPacket.requestValue string.
The newer design in reconfiguration will only accept InterfaceRequest,
so the above problem won't exist.

Fixed nodeID-to-int conversion by moving it inside
PaxosManager.handleIncomingPacket(.). The earlier design did it inside
PaxosPacketDemultiplexer but that will be circumvented if
handleIncomingPacket is called directly from outside as is currently
allowed by design.

Minor cleanup to various classes in gigapaxos.

Moved back SampleNodeConfig to nio.nioutils. That is where it must be.
It and all of nio have nothing to do with Auspice.



git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@1321 0b15c3b6-4604-4ca4-ad59-2e1af984e93c