Found a solution for MOB-358 (congestion during send which results in a timeout when adding a record)
Fixed two bugs related to the use of NIO:

1) One was that GnsMessenger (still used by NameServer) was calculating the size of data being sent incorrectly and returning a bogus value.
Which also meant that the retransmit part of the code in GnsMessenger was never working. GnsMessenger is probably going away anyway…

2) The contract that the send methods in NIO have is that they return one of 3 values indicating the status of the write.
0 meaning it might be sent at some point but hasn’t been yet, -1 meaning it can’t ever be sent or an int value equal to the size of the data written meaning that the
data will be sent. The third value is problematic (and was the case of a bug) because some methods alter the size of the data so they need to be careful to
return a value that their caller expects they will write not the amount they actually write. A better data representation for this return value might be a simple 
categorical like an enum that says what we actually mean. MightWrite, CantWrite, DidWrite for example. Nobody really cares about the actual amount written. Maybe we’ll do that.

All test currently pass without Gigipaxos. With Gigipaxos we're getting a timeout trying to add a record.