Working on MOB-362, MOB-363, MOB-364
Replicable now extends InterfaceReplicable. Reconfigurable now extends InterfaceReconfigurable. 
Made a first pass at implementing the necessary methods.
The getRequest method in InterfaceApplication is going to be on of the trickiest because it will require us to construct an arbitrary Packet instance from a JSONObject.

ReplicaControllerCoordinatorPaxos extends AbstractReplicaCoordinator and implements some of the methods. 
Still to be done: deleteReplicaGroup, deleteReplicaGroup, getReplicaGroup and getRequestTypes.

Nothing old is broken - tests pass.