Fixed a couple of issue with the latest commit.
Typo causing a bug on ConsistentReconfigurableNodeConfig.getValuesFromJSONArray where it was calling itself.

Added a two methods to ConsistentHashing and ConsistentReconfigurableNodeConfig.getValuesFromJSONArray that allow backward compatibility with the old app.
They are marked them as deprecated because they should be removed when we finish the transition to the new app.

Fixed a compiler error in DerbyPersistentReconfiguratorDB.getRandomInterfaceRequest. The String name parameter needs to be marked final 
so it can be accessible in the inner class.