Working on MOB-372: Integrate GNS with the gns.reconfiguration package
Implemented a handler for the new RequestActiveReplicas packets coming back to the LNS.

Also fixed an issue where the app wasn't properly handling DNS requests because the version (epoch)
number of the record in the database was wrong when the record was fresh.

All tests currently pass with the NullDemandProfile. WIll be testing next with  the default demand profile.