MOB-389 - InterfaceNodeConfig needs versioning with a getVersion() method.
Added a version number to InterfaceNodeConfig.
GNSNodeConfig implements this by supporting the reading of a version number as the first line of the hosts file.
GNSNodeConfig also periodically rereads the hosts file checking of a change in the number on the first line of the file if it exists. If the number changes the node config is updated.