More performance optimizations to gigapaxos' networking and serialization (currently the main bottleneck) that roughly double its capacity compared to the previous version.
Added a journal-based logger that is currently incomplete, i.e., there
is no support yet for periodic compaction or reading from the logs,
but it suffices to accurately capture the overhead of an efficient
custom logger (unlike derby DB), so it is useful for perfomance
instrumentation. We do also need a DB-like interface in order to
maintain group scalability, so the plan going forward is to
infrequently compact the journals into the current SQL logger.

The options DISABLE_LOGGING=false and ENABLE_JOURNALING=true are
needed to switch to journaling. Note that because there is currently
no mechanism to read from the journaled logs, it is effectively like
disabling logging, so the system can not recover correctly from the
failure of a majority of replicas.